refactored auth folder
