+++ castoredc_api/tests/test_import/fixtures_import.py
@@ -1,7 +1,7 @@
 import pytest
 
 from castoredc_api.study.castor_study import CastorStudy
-from auth import auth_data
+from castoredc_api.auth import auth_data
 
 
 @pytest.fixture(scope="session")
