+++ castoredc_api/tests/test_output/test_csv_output.py
@@ -1,6 +1,6 @@
 import pytest
 from csv_diff import load_csv, compare
-import auth.auth_data as auth_data
+from castoredc_api import auth as auth_data
 
 from castoredc_api.study.castor_study import CastorStudy
